issue 151 resolved

+++ development-codes/update.py
@@ -3,7 +3,6 @@ import os
 import yaml
 
 
-
 def make_html(current, parent, ancestors):
     
     catalog_dir = "https://raw.githubusercontent.com/kpegion/COLA-DATASETS-CATALOG/gh-pages/intake-catalogs/"
@@ -41,11 +40,16 @@ def update_json(tags, html_page, dataset_sub_name):
 
 
 
-def update_parents(ancestors):
+def make_ancestors(ancestors):
     #<!-- 1qaz2wsx -->
     
     ancestors = ancestors.split('/')
-    ancestors.remove('')
+
+    try:
+        ancestors.remove('')
+    except:
+        pass
+
     try:
         ancestors.remove('shared')
     except:
@@ -73,25 +77,47 @@ def update_parents(ancestors):
 
 
 
-
-    
-
-
-
 def catalog_parent( _path, _dataset_name):
 
-    
+    print(_dataset_name) 
     direct_parent = _path.split('/')[-2].lower()
     dict_file = [
-    {'description': 'COLDA '+ direct_parent + ' Data Caalog'},
+    {'description': 'COLDA '+ direct_parent.upper() + ' Data Caalog'},
     {'sources':{_dataset_name:[{'args':{'path':_path}},{'description':''},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
     }}]
-    
     direct_parent = direct_parent + '.yaml'
-    with open(direct_parent, 'w') as fp:
+    with open(direct_parent, 'a') as fp:
         yaml.dump(dict_file, fp)
+        print(direct_parent + " created")
 
-def update_links(child, dp_name):
+
+
+def catalog_parent_2( _path, _dataset_name, direct_parent):
+
+    direct_parent = direct_parent + '.yaml'
+    if not os.path.isfile(direct_parent):
+     
+        
+        dict_file = [
+        {'description': 'COLDA '+ direct_parent.upper() + ' Data Caalog'},
+        {'sources':{_dataset_name:[{'args':{'path':_path}},{'description':''},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
+        }}]
+
+    
+        with open(direct_parent, 'w') as fp:
+            yaml.dump(dict_file, fp)
+            print(direct_parent + " created")
+        
+    else:
+        
+        with open(direct_parent) as fp:
+            newdct = yaml.load(fp, Loader=yaml.FullLoader)
+
+        with open(direct_parent, 'w') as fp:
+            newdct[1]['sources'][_dataset_name+"tesdgdghdghtak"] =[{'args':{'path':_path}},{'description':'JadidTar'},{'driver':'intake.catalog.local.YAMLFileCatalog'},{'metadata':'{}'}]
+            yaml.dump(newdct, fp)
+
+def link_to_children(child, dp_name):
 
     searchee = "<!--qazwsxxswzaq-->"
     res = """               <a href="""+ "\""+ child+"\""+""" class="list-group-item">
